requirements.txt
langchain
langchain_groq
chromadb
streamlit
sentence_transformers
datasets
transformers
python-dotenv
pydantic
